Added `distribution` directory to build self-contained C++ and Python navground packages. It is meant to be used with `cpack` and `python setup.py bdist_wheel` to build installers for the different platforms.

Adapted packages'CMakeLists.txt files to support inclusion in top project. Added min required macOs version. Updated required cmake version. Added projects' version. Added a prefixed to target names. Added opt-out to build standalone executables or install Python files. Added support to bulld python extension in custom directory (to be specified by setup.py). Now uses project version when there are no git tags.

Restructured release workflow, splitting in several workflow to be called from main workflow.

Moved plugins to separate subdirectory in `navground_examples_cpp`.

Added minimal plugins packages as two packages with a minimal plugin skeleton for Python and C++.

`navground_{core|sim}_py`: added support to include DLL directories in Windows, so we don't need to install shared libraries inside the Python package anymore. Added exporting `NAVGROUND_DLL_PATH`.

Resolved/silenced all compile warning: removed unnecessary semicolons, added flags, corrected typos,

Moved `plugins.dsv` from `navground_core` to plugins but left default `NAVGROUND_PLUGINS_PREFIX` to `navground_core`. Now plugins packages can export the path where they are installed.

+++ navground_core/include/navground/core/buffer.h
@@ -146,8 +146,8 @@ inline size_t get_size(const BufferShape &shape) {
   if (!shape.size()) {
     return 0;
   }
-  return static_cast<size_t>(std::accumulate(std::begin(shape), std::end(shape),
-                                             1, std::multiplies<ssize_t>{}));
+  return std::accumulate(std::begin(shape), std::end(shape),
+                         static_cast<size_t>(1), std::multiplies<ssize_t>{});
 }
 
 /**
@@ -234,7 +234,7 @@ struct BufferDescription {
   std::vector<size_t> get_strides() const {
     std::vector<size_t> strides(shape.size());
     size_t size = get_scalar_size(type);
-    const int e = shape.size() - 1;
+    const int e = static_cast<int>(shape.size()) - 1;
     for (int i = 0; i <= e; ++i) {
       strides[e - i] = size;
       size *= shape[e - i];
